Zeroing in on switching the simulation engine
Moving the simulate call into the SimObject instead of the EventManager should let me start to send events to Queues in other objects instead of the main event loop.

There are a couple of classes that do not use the SimObject and directly include the EventManager (Intel and mc timers and the PacketQueue). I'm not overly concerned about the timers, but the packet queue could pose a challenge as I believe that is the base for all of the memory control.